move logprep.output.console_output to logprep.connector.console.output
